+++ examples/mlplatform/spark_k8s/main.tf
@@ -1,0 +1,36 @@
+locals {
+  spark_configuration = {
+    "spark.eventLog.dir"     = "s3a://spark-bucket"
+    "spark.eventLog.enabled" = "true"
+  }
+  spark_environment_variables = {
+    "ENV_VAR_1" : "env_var_1_value"
+    "ENV_VAR_2" : "env_var_2_value"
+  }
+}
+
+resource "vkcs_mlplatform_spark_k8s" "spark_k8s" {
+  name              = "tf-example"
+  availability_zone = "GZ1"
+  network_id        = vkcs_networking_network.app.id
+  subnet_id         = vkcs_networking_subnet.app.id
+
+  node_groups = [
+    {
+      node_count          = 2
+      flavor_id           = data.vkcs_compute_flavor.basic.id
+      autoscaling_enabled = true
+      min_nodes           = 2
+      max_nodes           = 100
+    }
+  ]
+  cluster_mode = "DEV"
+  registry_id  = vkcs_mlplatform_k8s_registry.k8s_registry.id
+  ip_pool      = data.vkcs_networking_network.extnet.id
+
+  suspend_after_inactive_min = 120
+  delete_after_inactive_min  = 1440
+
+  spark_configuration   = yamlencode(local.spark_configuration)
+  environment_variables = yamlencode(local.spark_environment_variables)
+}